Added SerializedPacket class, updated databuffer to use ushort for byte array lengths

+++ Chip.Net/Chip.Net/Data/DataBuffer.cs
@@ -175,7 +175,7 @@ namespace Chip.Net.Data
 		/// </summary>
 		/// <param name="source"></param>
 		public void Write(byte[] source) {
-			Write((short)source.Length);
+			Write((ushort)source.Length);
 			for (int i = 0; i < source.Length; i++)
 				Write((byte)source[i]);
 		}
@@ -222,7 +222,7 @@ namespace Chip.Net.Data
 		/// </summary>
 		/// <returns></returns>
 		public byte[] ReadByteArray() {
-			var length = ReadByte();
+			var length = ReadUInt16();
 			byte[] ret = new byte[length];
 			for (int i = 0; i < length; i++)
 				ret[i] = ReadByte();
